Refactor PrettyPrinter
- `PrettyPrinter` is now a proxy to the ReFrame's logger and takes
  particular care in formatting specially the progress output.
  It also prefixes errors and warnings with the program name.
- The general framework logger is now also capable of colorizing its
  output. This allows errors and warnings to be colorized when logged
  from anywhere in the framework.
