+++ reframe/core/logging.py
@@ -14,6 +14,7 @@ import reframe
 import reframe.core.debug as debug
 import reframe.utility.os_ext as os_ext
 from reframe.core.exceptions import ConfigError, LoggingError
+from reframe.utility import AnsiColorizer
 
 
 # Global configuration options for logging
@@ -352,6 +353,7 @@ class LoggerAdapter(logging.LoggerAdapter):
             }
         )
         self.check = check
+        self.colorize = False
 
     def __repr__(self):
         return debug.repr(self)
@@ -414,6 +416,19 @@ class LoggerAdapter(logging.LoggerAdapter):
 
     def verbose(self, message, *args, **kwargs):
         self.log(VERBOSE, message, *args, **kwargs)
+
+    def warning(self, message, *args, **kwargs):
+        if self.colorize:
+            message = AnsiColorizer.colorize(message, AnsiColorizer.yellow)
+
+        super().warning(message, *args, **kwargs)
+
+    def error(self, message, *args, **kwargs):
+        if self.colorize:
+            message = AnsiColorizer.colorize(message, AnsiColorizer.red)
+
+        super().error(message, *args, **kwargs)
+
 
 
 # A logger that doesn't log anything
